my_api/requirements.txt
aniso8601==9.0.1
asyncpg==0.27.0
blinker==1.7.0
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.89.0
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==2.0.1
h11==0.14.0
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2==2.9.9
psycopg2-binary==2.9.5
PyJWT==2.8.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
python-decouple==3.6
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7
rsa==4.9
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
Werkzeug==3.0.2
